+++ lib/app/server/plugins/apollo-error.plugin.d.ts
@@ -1,0 +1,3 @@
+import { ApolloServerPlugin } from '@rumbleship/apollo-server-hapi';
+import { RumbleshipContext } from './../../rumbleship-context/rumbleship-context';
+export declare const logErrorsPlugin: ApolloServerPlugin<RumbleshipContext>;