+++ src/app/index.ts
@@ -1,3 +1,2 @@
-export * from './config';
 export * from './rumbleship-context';
 export * from './server';