+++ src/app/server/plugins/hapi-plugins.d.ts
@@ -1,0 +1,2 @@
+declare module 'hapi-require-https';
+declare module 'hapi-request-id-header';