+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './app';
 export * from './gql';
 export * from './helpers';
 export * from './db';
+export * from '@rumbleship/apollo-server-hapi';