+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 
 ### Added
+  * Centralized `initServer` logic
 ### Removed
 ### Changed
+  * Extract convict config schemas to separate package, `@rumbleship/config`
 ### Fixed
 ### Deprecated
 ### Security